Changes: The name of the renaming file has been changed to "variable_maintenance". This file should now be used for any variable-related maintenance tasks. 	-"Old_names": This column holds the original names from the original API file.
	-"New_names": This column holds corrected/translated names.

	-"FAC_Aggregation": This column is for the grouping of similar variables. If no group is specified (blank or null), the value from the "New_names" column will be used as the group.

	-"Include_Variable": This column indicates if a variable needs to be included (1) or not (0). The manual elimination of variables in the code has been removed.

With these changes, there is no longer a need for the user to manually maintain variable-related elements in the code, so those lines were removed from the script.

Variable aggregation is performed at the beginning of the code, allowing for a natural flow throughout the rest of the script. Note that this can be easily modified by altering the name variable used from the maintenance file.

A "variables_to_check.xlsx" file is generated, containing all variables not present in the maintenance file. If there are any rows in this file at the end of the code, a warning is generated.

The removal of variables with fewer than 5 POIs has been removed.